requirements.txt
dnspython
flask
flask_cors
numpy
pymongo
python-dotenv
opencv-python
imutils